Added reverse square root

+++ square_root/c.c
@@ -1,11 +1,11 @@
-uint8_t square_root(uint8_t num) {
-    uint8_t g = num / 2;
+float square_root(float num) {
+    float g = num / 2;
 
-    for(uint8_t i = 0; i < 10; i++) {
+    for(float i = 0; i < 10; i++) {
         g = 0.5* (g + num / g);
     }
 
     return g;
 }
 
-uint8_t square_root_min(uint8_t n) { uint8_t g = n / 2; for(uint8_t i = 0; i < 10; i++) { g = 0.5 * (g + n / g); } return g; }
+float square_root_min(float n) { float g = n / 2; for(float i = 0; i < 10; i++) { g = 0.5 * (g + n / g); } return g; }
